move make_filter_handles_op to transform_ext/utils/
